[CHORE] moved TestHTTPRouteAttribute to test subfolder - instrumentation/github.com/labstack/echo/otelecho/test
