+++ instrumentation/github.com/labstack/echo/otelecho/test/echo_test.go
@@ -256,3 +256,37 @@ func TestSpanNameFormatter(t *testing.T) {
 		})
 	}
 }
+
+func TestHTTPRouteAttribute(t *testing.T) {
+	// Set up an in-memory span recorder
+	sr := tracetest.NewSpanRecorder()
+	tp := trace.NewTracerProvider(trace.WithSpanProcessor(sr))
+
+	// Create Echo instance with middleware
+	e := echo.New()
+	e.Use(otelecho.Middleware("test-service", otelecho.WithTracerProvider(tp)))
+	e.GET("/users/:id", func(c echo.Context) error {
+		return c.String(http.StatusOK, "ok")
+	})
+
+	// Simulate a request
+	r := httptest.NewRequest("GET", "/users/123", nil)
+	w := httptest.NewRecorder()
+	e.ServeHTTP(w, r)
+
+	// Check response
+	assert.Equal(t, http.StatusOK, w.Result().StatusCode)
+
+	// Verify span attributes
+	spans := sr.Ended()
+	assert.Len(t, spans, 1, "expected one span")
+
+	found := false
+	for _, attr := range spans[0].Attributes() {
+		if attr.Key == "http.route" && attr.Value.AsString() == "/users/:id" {
+			found = true
+			break
+		}
+	}
+	assert.True(t, found, "http.route attribute not found or incorrect, got %v", spans[0].Attributes())
+}
